apply review, add docs, separate iterator.go
