+++ immutable_tree.go
@@ -223,71 +223,3 @@ func (t *ImmutableTree) nodeSize() int {
 	})
 	return size
 }
-
-type Iterator struct {
-	start, end []byte
-
-	key, value []byte
-
-	valid bool
-
-	// next is the argument for retrieving the sequence starting from the next element.
-	// By calling next manually, Iterator can control the execution flow.
-	t *traversal
-}
-
-func (t *ImmutableTree) Iterator(start, end []byte, ascending bool) *Iterator {
-	iter := &Iterator{
-		start: start,
-		end:   end,
-		valid: true,
-		t:     t.root.newTraversal(t, start, end, ascending, false, false),
-	}
-
-	iter.Next()
-	return iter
-}
-
-var _ dbm.Iterator = &Iterator{}
-
-func (iter *Iterator) Domain() ([]byte, []byte) {
-	return iter.start, iter.end
-}
-
-func (iter *Iterator) Valid() bool {
-	return iter.valid
-}
-
-func (iter *Iterator) Key() []byte {
-	return iter.key
-}
-
-func (iter *Iterator) Value() []byte {
-	return iter.value
-}
-
-func (iter *Iterator) Next() {
-	node := iter.t.next()
-	if node == nil {
-		iter.t = nil
-		iter.valid = false
-		return
-	}
-
-	if node.height == 0 {
-		iter.key, iter.value = node.key, node.value
-		return
-	}
-
-	iter.Next()
-}
-
-func (iter *Iterator) Close() error {
-	iter.t = nil
-	iter.valid = false
-	return nil
-}
-
-func (iter *Iterator) Error() error {
-	return nil
-}
